requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
distlib==0.3.6
Django==3.1.4
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.11.0
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.2.0
promise==2.3
PyJWT==1.7.1
pytz==2020.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tzdata==2023.3
Unidecode==1.1.1
virtualenv==20.21.0
virtualenv-clone==0.5.7
